requirements.txt
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==2.2.2
django-cors-headers==3.11.0
djangorestframework==3.11.2
factory-boy==3.2.1
Faker==12.3.3
idna==3.3
iniconfig==1.1.1
packaging==21.3
petl==1.7.8
pluggy==1.0.0
psycopg2-binary==2.8.6
py==1.11.0
pyparsing==3.0.7
pytest==7.0.1
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.8
